feat(tasks): the delivery is ONE component, and the session aside draws it
The session's Task tab was a summary card — the delivery's name, its rollup and
three buttons. Everything that makes a delivery worth opening (its description,
its parts, the sessions filed under each, the comments and their attachments,
the pull requests, what blocks it, its status, its claim) lived on a page you
had to leave the session to reach, and half of it could not be reached from
there at all.

Reproducing it in the aside would have been a second implementation of every one
of those gestures. So the body of `/tasks/:id` moved into `DeliveryDetail` and
BOTH surfaces render it. It is not a mirror anyone has to keep in sync — it is
the same component over the same `/api/tasks`, which is what makes "change it
here and it changes on the board" true by construction. `dense` is the only
difference and it is LAYOUT ONLY: the aside is a ~380px column, so the work|facts
split becomes one column and the rail follows the tabs. Nothing is dropped — a
control that exists on the page and not in the aside is the asymmetry this
move exists to end.

STOPPING A SESSION NOW OFFERS TO FINISH THE PART, not the delivery. A session is
filed under a SUBTASK — the unit of work, the thing it actually did, and the only
thing one session can honestly close. The delivery follows its parts: it is
marked delivered ONLY when this was the last open one, and the button says so
before it is pressed. A stop that silently closed a whole delivery because it
happened to be the last part is a surprise; the same act announced beforehand is
a decision. A session whose part cannot be resolved keeps the plain confirmation
rather than offering to finish something the dialog cannot name.

NATIVE `<select>`s ARE GONE from the board. "Pick a task…" in the blocked-by rail
and the composer's part chooser both drew the OS menu: it ignores this palette in
both themes, it cannot search, and on a phone it is the one control there that
misses the 44px target. Both now use `Select` from the settings primitives — the
same control every settings screen uses, searchable exactly when a board has
grown past the point where scrolling works.

THE MCP KNOWS THE RULE. `agentistics_task_session` filed at the delivery level,
which the server now refuses, so the tool was broken by the previous commit and
would have answered a bare `{ok:false}`. It takes `subtaskId`, refuses in a
SENTENCE naming the two calls that get the caller unstuck, and
`agentistics_task_subtask` no longer advertises the one-session `sessionId`
field — a subtask holds any number, and the link lives on the session.

Also: the delivery's tabs and the blocked-by rows speak the reader's language,
and `copy.test.ts` now WALKS the record instead of skipping `status` by name —
that skip is why `tabs` arrived unchecked.
