+++ packages/web/src/components/tasks/SessionTasksTab.tsx
@@ -1,27 +1,29 @@
 /**
- * SessionTasksTab — what this session is FILED UNDER, and the form to file it somewhere new.
+ * SessionTasksTab — the DELIVERY this session belongs to, whole, beside the conversation.
  *
- * The session workspace could tell you a session belonged to a task (a badge on the row) and gave
- * you nowhere to act on it: unfiling meant hunting through a menu, and creating a task from the
- * session you were sitting in meant leaving it for the board. So the tab is both — the task's own
- * numbers, the way out, and the SAME composer the board's "New task" opens, with this session
- * already linked.
+ * It used to be a summary card: the delivery's name, its rollup, and three buttons. Everything that
+ * makes a delivery worth opening — its description, its parts, the sessions filed under each, the
+ * comments and their attachments, the pull requests, what is blocking it, its status, its claim —
+ * was on a page you had to leave the session to reach, and half of it could not be reached from
+ * here at all.
  *
- * It shows the task's rollup rather than repeating the session's: the point of filing a session is
- * that the TASK becomes measurable, and the tab is where you see that it worked.
+ * So this tab now renders `DeliveryDetail`, the very component `/tasks/:id` renders, in its `dense`
+ * one-column shape. **It is not a mirror that has to be kept in sync — it is the same component
+ * over the same `/api/tasks`**, which is what makes "change it here and it changes on the board"
+ * true by construction rather than by maintenance.
+ *
+ * What stays local is the FILING: which delivery this session belongs to is a question about the
+ * SESSION, and it is the one thing this tab answers that the page does not.
  */
 
 import { useMemo, useState } from 'react'
 import { ExternalLink, Link2, Plus, Unlink } from 'lucide-react'
 import { useIsMobile } from '../../hooks/useIsMobile'
-import {
-  NA, STATUS, button, fmtInt, microLabel, numeric, pill, surface, type BoardStatus,
-} from './board'
-import { useMoney } from './money'
-import { TaskProgressBar } from './TaskProgressBar'
+import { button, microLabel, pill, surface } from './board'
 import { boardCopy } from './copy'
 import { TaskComposer } from './TaskComposer'
-import { attachSession, detachSession, useTaskDetail, useTaskList } from '../../lib/tasks'
+import { detachSession, useTaskDetail, useTaskList } from '../../lib/tasks'
+import { DeliveryDetail } from './DeliveryDetail'
 import { SessionFiling } from './SessionFiling'
 import { BetaTag } from '../BetaTag'
 
@@ -36,7 +38,6 @@ export interface SessionTasksTabProps {
 
 export function SessionTasksTab(p: SessionTasksTabProps) {
   const isMobile = useIsMobile()
-  const money = useMoney()
   const pt = p.lang === 'pt'
   const { rows, reload } = useTaskList()
   const [composing, setComposing] = useState(false)
@@ -64,12 +65,17 @@ export function SessionTasksTab(p: SessionTasksTabProps) {
 
   const refresh = async () => { await reload(); await reloadDetail(); p.onChanged?.() }
 
-  const link = async (taskId: string) => {
-    setBusy(true)
-    await attachSession(taskId, p.session.id)
-    setBusy(false)
-    await refresh()
-  }
+  /**
+   * The PART this session sits in, by name.
+   *
+   * Read off the delivery's own session rows rather than from anything the fleet carries: the
+   * fleet knows the delivery (that is the badge on the row) and nothing about which subtask, and
+   * inventing one from the first subtask would name a part nobody chose.
+   */
+  const here = useMemo(() => {
+    const sub = detail?.sessions.find(r => r.id === p.session.id)?.subtaskId
+    return sub ? detail?.subtasks.find(t => t.id === sub)?.title : undefined
+  }, [detail, p.session.id])
 
   const unlink = async () => {
     setBusy(true)
@@ -88,75 +94,69 @@ export function SessionTasksTab(p: SessionTasksTabProps) {
       </div>
       {p.session.task
         ? (
-          <div style={{ ...surface, padding: 12, display: 'grid', gap: 10 }}>
-            <div style={{ display: 'flex', alignItems: 'center', gap: 8 }}>
-              <span style={microLabel}>{pt ? 'Esta sessão está em' : 'This session is filed under'}</span>
-              <span style={{ flex: 1 }} />
-              {current && (
-                <span style={pill(STATUS[current.task.status as BoardStatus]?.color)}>
-                  {STATUS[current.task.status as BoardStatus]?.label ?? current.task.status}
-                </span>
-              )}
-            </div>
-            <div style={{ fontSize: 13.5, fontWeight: 650, color: 'var(--text-primary)' }}>
-              {p.session.task}
-            </div>
-
-            {current && (
-              <>
-                <TaskProgressBar
-                  done={current.counts.subtasksDone}
-                  total={current.counts.subtasks}
-                />
-                {/* The TASK's numbers, not this session's. Filing a session is what makes the task
-                    measurable, and this is where you see that it worked. */}
-                <div style={{ display: 'flex', gap: 16, flexWrap: 'wrap' }}>
-                  {([
-                    [pt ? 'Custo' : 'Cost', money(current.rollup.costUSD)],
-                    [pt ? 'Rodadas' : 'Rounds', fmtInt(current.rollup.rounds)],
-                    [pt ? 'Sessões' : 'Sessions', String(current.rollup.sessionsUsed)],
-                  ] as const).map(([label, value]) => (
-                    <span key={label}>
-                      <span style={{ ...microLabel, display: 'block' }}>{label}</span>
-                      <span style={{
-                        ...numeric, display: 'block',
-                        color: value === NA ? 'var(--text-tertiary)' : 'var(--text-secondary)',
-                      }}>{value}</span>
-                    </span>
-                  ))}
-                </div>
-              </>
-            )}
-
-            <div style={{ display: 'flex', gap: 8, flexWrap: 'wrap' }}>
-              {current && p.onOpenTask && (
+          <>
+            {/* The filing bar — the one question that is about the SESSION rather than about the
+                delivery, kept above the delivery itself so the two are never confused. */}
+            <div style={{ ...surface, padding: 10, display: 'grid', gap: 8 }}>
+              <div style={{ display: 'flex', alignItems: 'center', gap: 8, flexWrap: 'wrap' }}>
+                <span style={microLabel}>{pt ? 'Esta sessão está em' : 'This session is filed under'}</span>
+                <span style={{ flex: 1 }} />
+                {here && (
+                  <span style={{ ...pill(), fontSize: 10 }}>
+                    {pt ? 'na parte' : 'in'} {here}
+                  </span>
+                )}
+              </div>
+              <div style={{ display: 'flex', gap: 8, flexWrap: 'wrap' }}>
+                {current && p.onOpenTask && (
+                  <button
+                    style={{ ...button(isMobile), height: isMobile ? 44 : 26 }} disabled={busy}
+                    onClick={() => p.onOpenTask?.(current.task.id)}
+                  >
+                    <ExternalLink size={13} /> {pt ? 'Abrir no board' : 'Open on the board'}
+                  </button>
+                )}
                 <button
-                  style={button(isMobile)} disabled={busy}
-                  onClick={() => p.onOpenTask?.(current.task.id)}
+                  style={{ ...button(isMobile), height: isMobile ? 44 : 26 }}
+                  disabled={busy} onClick={() => setPicking(true)}
                 >
-                  <ExternalLink size={13} /> {pt ? 'Abrir a tarefa' : 'Open the task'}
+                  <Link2 size={13} /> {pt ? 'Trocar' : 'Change'}
                 </button>
-              )}
-              <button style={button(isMobile)} disabled={busy} onClick={() => setPicking(true)}>
-                <Link2 size={13} /> {pt ? 'Trocar' : 'Change'}
-              </button>
-              <button
-                style={{ ...button(isMobile), color: 'var(--accent-red)' }}
-                disabled={busy}
-                onClick={() => void unlink()}
-              >
-                <Unlink size={13} /> {pt ? 'Desvincular' : 'Unfile'}
-              </button>
+                <button
+                  style={{ ...button(isMobile), height: isMobile ? 44 : 26, color: 'var(--accent-red)' }}
+                  disabled={busy}
+                  onClick={() => void unlink()}
+                >
+                  <Unlink size={13} /> {pt ? 'Desvincular' : 'Unfile'}
+                </button>
+              </div>
             </div>
-          </div>
+
+            {/* THE DELIVERY ITSELF — the same component the board's own page draws. */}
+            {detail
+              ? (
+                <DeliveryDetail
+                  id={detail.task.id}
+                  detail={detail}
+                  lang={p.lang}
+                  reload={refresh}
+                  dense
+                />
+              )
+              : (
+                <div style={{ ...surface, padding: 12, fontSize: 12, color: 'var(--text-tertiary)' }}>
+                  {pt ? 'Carregando a entrega…' : 'Loading the delivery…'}
+                </div>
+              )}
+          </>
         )
         : (
           <div style={{ ...surface, padding: 12, display: 'grid', gap: 9 }}>
-            <span style={microLabel}>{pt ? 'Sem tarefa' : 'Not filed under a task'}</span>
+            <span style={microLabel}>{pt ? 'Sem entrega' : 'Not filed under a delivery'}</span>
             <span style={{ fontSize: 12, color: 'var(--text-tertiary)', lineHeight: 1.5 }}>
               {pt
-                ? 'Uma sessão fora de uma tarefa continua sendo medida — o que não existe é o custo POR entrega. Vincule esta a uma tarefa, ou crie uma aqui mesmo.'
-                : 'A session outside a task is still measured — what does not exist is the cost PER delivery. File this one under a task, or create one right here.'}
+                ? 'Uma sessão fora de uma entrega continua sendo medida — o que não existe é o custo POR entrega. Filie esta a uma subtarefa, ou crie a entrega aqui mesmo.'
+                : 'A session outside a delivery is still measured — what does not exist is the cost PER delivery. File this one under a subtask, or create the delivery right here.'}
             </span>
             <div style={{ display: 'flex', gap: 8, flexWrap: 'wrap' }}>
               <button style={button(isMobile)} disabled={busy} onClick={() => setPicking(true)}>
@@ -165,7 +165,6 @@ export function SessionTasksTab(p: SessionTasksTabProps) {
             </div>
           </div>
         )}
-
 
       {/* The composer, INLINE — the same form the board opens in a dialog, with this session
           pre-linked. Creating a task from the session you are sitting in should not mean leaving
